Variants of PSD2 in the hope to improve accuracy
With global variable variant_with_background_fluxes == True, the code
accounts for the possibility that individuals of a species may enter a
patch while waiting, fail to establish and yet disperse to a
neighbouring patch from there.  As a result, patches in waiting state
don't act as a block to dispersal.

With global variable variant_with_P_to_D_transitions == True, it
accounts for the possibility of direct transitions from P to D states.
This can happen for example when turnover rates are high and local
growth rates change fast.